Reactivated istanbul coverage reports.
